# service/apigateway/docs/swagger.yaml
definitions:
  errors.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      retryable:
        type: boolean
      status:
        type: string
      trace_id:
        type: string
      type:
        $ref: '#/definitions/errors.ErrorType'
      validations:
        items:
          $ref: '#/definitions/errors.ValidationError'
        type: array
    type: object
  errors.ErrorType:
    enum:
    - INTERNAL
    - NOT_FOUND
    - BAD_REQUEST
    - CONFLICT
    - UNAUTHORIZED
    - FORBIDDEN
    - TIMEOUT
    - UNAVAILABLE
    type: string
    x-enum-varnames:
    - ErrorTypeInternal
    - ErrorTypeNotFound
    - ErrorTypeBadRequest
    - ErrorTypeConflict
    - ErrorTypeUnauthorized
    - ErrorTypeForbidden
    - ErrorTypeTimeout
    - ErrorTypeUnavailable
  errors.ValidationError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
...
  response.ApiResponseSlider:
    properties:
      data:
        $ref: '#/definitions/response.SliderResponse'
      message:
        type: string
      status:
        type: string
    type: object
  response.ApiResponseSliderAll:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  response.ApiResponseSliderDelete:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  response.ApiResponseSliderDeleteAt:
    properties:
      data:
        $ref: '#/definitions/response.SliderResponseDeleteAt'
      message:
        type: string
      status:
        type: string
    type: object
...
  response.SliderResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  response.SliderResponseDeleteAt:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/slider-command/create:
    post:
      consumes:
      - multipart/form-data
      description: Create a new slider with an image upload
      parameters:
      - description: Slider name
        in: formData
        name: name
        required: true
        type: string
      - description: Slider image
        in: formData
        name: image_slider
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created slider
          schema:
            $ref: '#/definitions/response.ApiResponseSlider'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Failed to create slider
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Create slider
      tags:
      - Slider Command
  /api/slider-command/permanent/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently delete a slider record by its ID
      parameters:
      - description: Slider ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted slider record permanently
          schema:
            $ref: '#/definitions/response.ApiResponseSliderDelete'
        "400":
          description: Invalid slider ID
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Failed to delete slider permanently
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Permanently delete slider
      tags:
      - Slider Command
  /api/slider-command/permanent/all:
    post:
      consumes:
      - application/json
      description: Permanently delete all trashed slider records
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted all sliders permanently
          schema:
            $ref: '#/definitions/response.ApiResponseSliderAll'
        "500":
          description: Failed to delete sliders permanently
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Permanently delete all trashed sliders
      tags:
      - Slider Command
  /api/slider-command/restore/{id}:
    post:
      consumes:
      - application/json
      description: Restore a trashed slider record by its ID
      parameters:
      - description: Slider ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully restored slider
          schema:
            $ref: '#/definitions/response.ApiResponseSliderDeleteAt'
        "400":
          description: Invalid slider ID
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Failed to restore slider
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Restore trashed slider
      tags:
      - Slider Command
  /api/slider-command/restore/all:
    post:
      consumes:
      - application/json
      description: Restore all trashed slider records
      produces:
      - application/json
      responses:
        "200":
          description: Successfully restored all sliders
          schema:
            $ref: '#/definitions/response.ApiResponseSliderAll'
        "500":
          description: Failed to restore sliders
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Restore all trashed sliders
      tags:
      - Slider Command
  /api/slider-command/trashed/{id}:
    post:
      consumes:
      - application/json
      description: Move a slider record to trash by its ID
      parameters:
      - description: Slider ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully moved slider to trash
          schema:
            $ref: '#/definitions/response.ApiResponseSliderDeleteAt'
        "400":
          description: Invalid slider ID
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Failed to move slider to trash
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Move slider to trash
      tags:
      - Slider Command
  /api/slider-command/update/{id}:
    post:
      consumes:
      - multipart/form-data
      description: Update an existing slider's name or image
      parameters:
      - description: Slider ID
        in: path
        name: id
        required: true
        type: integer
      - description: Slider name
        in: formData
        name: name
        type: string
      - description: Slider image
        in: formData
        name: image_slider
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated slider
          schema:
            $ref: '#/definitions/response.ApiResponseSlider'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Failed to update slider
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Update slider
      tags:
      - Slider Command
...
swagger: "2.0"